+++ src/LionWeb-CSharp/core/M1/M1Extensions.cs
@@ -89,6 +89,43 @@ public static class M1Extensions
     }
 
     /// <summary>
+    /// Returns all the preceding siblings of the base node <param name="self"></param>
+    /// Optionally includes <paramref name="self"/>.
+    /// 
+    /// Starting from index 0 to the index of base node, includes the index of <param name="self"></param>
+    /// if <paramref name="includeSelf"/> is set to true.  
+    /// </summary>
+    /// <param name="self">Base node to find preceding siblings of.</param>
+    /// <param name="includeSelf">If true, the result includes <paramref name="self"/>.</param>
+    /// <returns>List of preceding siblings of <param name="self"></param>.</returns>
+    /// <exception cref="TreeShapeException">If <paramref name="self"/> has no parent, or is not a child in a multiple containment.</exception>
+    public static IEnumerable<INode> PrecedingSiblings(this INode self, bool includeSelf = false)
+    {
+        List<INode> list = GetContainmentNodes(self);
+        var takeCount = includeSelf ? list.IndexOf(self) + 1 : list.IndexOf(self);
+        return list.Take(takeCount);
+    }
+
+    /// <summary>
+    /// Returns all the following siblings of the base node <param name="self"></param>
+    /// Optionally includes <paramref name="self"/>.
+    ///
+    /// Starting from the following index of <param name="self"></param>, includes the index of <paramref name="self"/>
+    /// if <paramref name="includeSelf"/> is set to true.  
+    /// </summary>
+    ///
+    /// <param name="self">Base node to find following siblings of.</param>
+    /// <param name="includeSelf">If true, the result includes <paramref name="self"/>.</param>
+    /// <returns>List of following siblings of <param name="self"></param>.</returns>
+    /// <exception cref="TreeShapeException">If <paramref name="self"/> has no parent, or is not a child in a multiple containment.</exception>
+    public static IEnumerable<INode> FollowingSiblings(this INode self, bool includeSelf = false)
+    {
+        List<INode> list = GetContainmentNodes(self);
+        var skipCount = includeSelf ? list.IndexOf(self) : list.IndexOf(self) + 1;
+        return list.Skip(skipCount);
+    }
+
+    /// <summary>
     /// Replaces <paramref name="self"/> in its parent with <paramref name="replacement"/>.
     ///
     /// Does <i>not</i> change references to <paramref name="self"/>.
@@ -154,6 +191,20 @@ public static class M1Extensions
         bool includeAnnotations = false) =>
         Children<INode>(self, includeSelf, includeAnnotations);
 
+
+    /// <summary>
+    /// Returns the first ancestor of <paramref name="self"/> that matches with type <typeparamref name="T"/>
+    /// Optionally includes <paramref name="self"/>.
+    /// </summary>
+    /// <param name="self">Base node to find ancestor of.</param>
+    /// <param name="includeSelf">If true, the result includes <paramref name="self"/>.</param>
+    /// <typeparam name="T"></typeparam>
+    /// <returns>The first ancestor of <paramref name="self"/> that matches with type <typeparamref name="T"/> or <c>null</c> </returns>
+    public static T? Ancestor<T>(this INode self, bool includeSelf = false) where T : INode =>
+        self.Ancestors(includeSelf)
+            .OfType<T>()
+            .FirstOrDefault();
+
     /// <summary>
     /// Enumerates all direct and indirect parents of <paramref name="self"/>.
     /// Optionally includes <paramref name="self"/>.
@@ -163,6 +214,24 @@ public static class M1Extensions
     /// <returns>All direct and indirect parents of <paramref name="self"/>.</returns>
     public static IEnumerable<INode> Ancestors(this INode self, bool includeSelf = false) =>
         Ancestors<INode>(self, includeSelf);
+
+    private static List<INode> GetContainmentNodes(INode self)
+    {
+        INode? parent = self.GetParent();
+        if (parent == null)
+            throw new TreeShapeException(self, "Cannot get siblings of a node with no parent");
+
+        Containment containment = parent.GetContainmentOf(self)!;
+        if (!containment.Multiple)
+            throw new TreeShapeException(self, "A single containment does not have siblings");
+
+        var value = parent.Get(containment);
+        if (value is not IEnumerable enumerable)
+            // should not happen
+            throw new TreeShapeException(self, "A single containment does not have siblings");
+
+        return [..containment.AsNodes<INode>(enumerable)];
+    }
 
     internal static IEnumerable<T> Descendants<T>(T self, bool includeSelf = false,
         bool includeAnnotations = false) where T : class, IReadableNode